+++ src/internals/TooltipSet/index.ts
@@ -1,0 +1,2 @@
+export { TooltipSet } from './TooltipSet';
+export type { TooltipSetProps } from './TooltipSet';