+++ src/internals/index.ts
@@ -2,3 +2,4 @@ export * from './CodeWrapper';
 export * from './ColorItem';
 export * from './ColorPalette';
 export * from './IconSet';
+export * from './TooltipSet';